Add a relationship of Release to Task
Given the way that we've operated so far, a Release contains many Tasks
(release requests for multiple apps). This commit keeps true to this
workflow by allowing multiple Tasks (app deploys) to be bundled up with
a single Release.

This involved removing the old 'release' field from Task and creating
the relevant association using a reference to Release (migration).