Remove TasksController
As tasks are now managed entirely within a release,
this controller is no longer necessary.

Perhaps, in the future, there may arise a need for
a controller to do a subset of actions on a task
from within a release (eg mark as done, delete),
but in the meantime, I think we should remove this.