Add shortname attribute on Application
An application has a "shortname" which will be used to determine the name of
its deploy metrics in graphite. This should correspond to the name of the
govuk::app type defined in Puppet.

This commit adds a "shortname" field which defaults to the repository name (i.e.
"bar" if "repo" is "foo/bar"), but can also be set explicitly.