Add release candidate announcer
This is designed to announce the release candidate of a particular app
in campfire. The release candidate shows the diff between the last
deployed version and the newest passing build.

The newest passing build is calculated by looking at the release branch
and seeing which tag corresponds to the same sha.