Show the list of unreleased commits, if available.
Instead of just displaying the last page of commits from the GitHub API,
we can use the comparison API to get all unreleased commits. This is
useful for apps that tend to end up with large pull requests, or which
end up getting batches of features released at once; in either case, the
latest released commit may not be visible in the commit log.