Name API applications correctly.
When we create a new app and deploy it for the first time, the release
app will have a stab at reconstructing the app's name from the address
of its repository. Unfortunately, Rails doesn't know that "API" is an
acronym, rather than a normal word, so whenever we deploy new APIs we
end up with applications called "Manuals Api" and "Finder Api" and other
such incorrectitude, until someone excessively pedantic goes in and
corrects the capitalisation manually.