Add a list of recent deployments to each app
This list can be useful when supporting apps and diagnosing problems, so
we can see the history of what has been deployed and when.

For some apps (Whitehall, for example), there will be several pull
requests merged before anything gets deployed to a non-preview
environment, so this filters out the automated deployments of the
`release` branch to the preview environment.

A possible future feature is to spot when an app has been deployed to
staging and then to production shortly afterwards, as this is the way we
usually release code. However, hopefully this is still enough
functionality to be useful.