Bump uglifier and tinder gem versions
The asset precompile was failing in production due to the following
error:

    NameError: uninitialized constant Uglifier::VERSION

Other reports of this error were due to an incompatibility between a
version of uglifier and a version of sprockets. They were always solved
by upgrading the version of uglifier to one which works with newer
versions of sprokets. Bumping to the latest version of Uglifier
introduced a dependency clash with the version of json required by
tinder. Bumping tinder a minor version changed the requirements of
tinder to use the newer version of json.