Update deploy job URLs
The host changed as part of the migration. The old ones are being redirected, but it's better to use the new one in the first place.

Also the name of the Production job changed, it's now consistent with the Staging job, which is good, but meant the deploy link wasn't working.