Bump Ruby to 2.2.3
Bumps Ruby and the `mysql2` gem for compatibity with 2.2.3. The `mysql2`
version is the latest Rails compatible version.

I had some troubles building `capybara-webkit` so replaced it with the
more modern `poltergeist` which is generally faster, more stable and
doesn't rely on QT which is generally considered as the framework
equivalent of pure hell. I also bumped `capybara` since `poltergeist`
was also at latest.