Show application status notes on the show and deploy pages
Notes are currently only surfaced on the index page, so if an important
message like "don't deploy without talking to team_x" is entered, the
user doing a deploy may not notice it.  By putting the note in a warning
alert on the show and deploy pages it should be more obvious to the user
that they should take care.

This may act as a way of reducing the need for the badger to act as a
physical cork board for these sorts of notice.  Which is prone to notices
falling off.