Be more consistent with flash types
We use "alert", "error", and "notice" in the app, but where we're using
"alert" it seems more appropriate to use "error".  Also, these flash
types aren't supported by bootstrap or govuk_admin_template as alert
styles, so the messages do not stand out in the template.  We avoid this
by translating from our alert names to the bootstrap ones in the CSS.