Remove check of deployment dashboard and add link to all apps
The Grafana API call which checks whether an app has a deployment
dashboard currently fails on non-development environments because
hostnames like `grafana.publishing.service.gov.uk` are not available on
integration/staging/production.

We plan to update the hostnames and firewall rules so that we can
reinstate this check, but in the meantime it is more useful to add a
link for every app since the majority of them do have a dashboard.