Introduce Site object to contain global status notes
This is a fairly naive singleton object.  It probably won't stand up to
a race condition with multiple people trying to create things at the same
time, but we won't present a UI that allows this.  Rather than naming this
for the status note feature on it's own we've used a more generic name.
The intention here is that we could extend this object later to store
other global site options that we want to persist in the database.