Make deploy page render when no production deploy is present
The controller action has a conditional that stops it generating the list
of commits for comparison when no production deploy is present, but the
view always assumes a production deploy is present and will break with
`NoMethodError` when trying to call `version` on `nil`.  It's likely that
in this case we'd actually want to fetch a different set of commits to
show the diff between nothing and the current tag, but this is an easier
fix.