Provide an accessor on Site to fetch the instance
It's another naive singleton implementation.  This method doesn't attempt
to memoize, so you should locally store the instance if you want to access
it multiple times.  We don't memoize because otherwise the instance
wouldn't update with changes until the Site class itself was reloaded (e.g
by a deploy, or new server worker spinning up).  This would be pretty
dangerous.