Add UI on editing the Site instance
We use a singular resource controller because we only ever have one
instance of Site to edit.  We also provide a controller/view level helper
method for memoizing the `Site.settings` call on a per-request basis and
it's the instance returned by this that we use in the views and
controllers.