Add app deployments in separate tab
This moves the application deployments into a separate page. To do so
we add tab navigation, which we also use to link to the edit page
(which is a button now).

This is part of a spike into using new admin patterns on GOV.UK. This
change will make it easier to experiment with the new patterns.

https://trello.com/c/09fVPrRN/64-try-to-convert-an-existing-admin-app-to
-govuk-frontend-component-system