Simplify forms in applications
This simplifies the form usage in the app.

- Use Simple form instead of Formtastic. The govuk_admin_template gem
already has a Simple form config, which uses bootstrap. This means we
can remove the `formtastic-bootstrap` gem as well.
- Remove the typeahead. This is only used by the form to submit new
deployments, and that happens so seldomly that it's not worth having
the JS dependencies for that.
- Remove the date picker. This is also just used by the deployment
form. By removing it we get rid of a whole range of dependencies.

Updating to simple form prepares this app for the new GOV.UK Frontend
based layout (https://github.com/alphagov/release/pull/228).

https://trello.com/c/09fVPrRN