Expose the previous version of a deploy
`Deployment#previous_deployment` is not relative to the deployment,
which the name seems to imply (otherwise it should have been a class
method). This is fine because we currently only use it to show the
previous deployment of the current deployment.

This change will make the method return the actual previous deployment,
so that you could do
`Deployment.last.previous_deployment.previous_deployment` and get the
3rd newest deployment, instead of just the 2nd newest deployment.