Change "In sync?" to Status in the applications table
This changes the status of applications to be reported as one of 3
states: production and staging differ, integration differs from
production or staging, and all environments match.

Red now goes back to representing differences between production and
staging, and an orange status cell is used to indicate the integration
differences.

This should hopefully make it clearer what state an application is in.