Remove special deploy permissions
We've had a few cases where it was unclear how to deploy correctly,
because someone didn't have the `deploy` permission in the Release app.
This hides the deploy buttons on the release screen.

This PR removes the permission and make everything available to all
users. This is fine because there's not much you can actually do
without production access (the deploy link just links out to Jenkins).
It simplifies the app and cuts down on things people have to do once
they have prod access.