Simplify application notes
This removes the ability to edit application status notes via a modal,
linked from the dashboard. Instead, it links to the edit page where the
same functionality exists.

This removes a lot of Javascript and dependency on Bootstrap, which
will be beneficial if we want to convert this application to use the
[component
system](https://govuk-publishing-components.herokuapp.com/component-guid
e/layout_for_admin).

https://trello.com/c/o3RVVD1w