Always show links to dashboards
This effectively reverts [1]. Currently we have an issue where
dashboards have to exist in Carrenza, so that the Release app thinks
it exists, and links to the dashboard in AWS. This "check if the
dashboard exists" feature also blocks removing redundant dashboards
from Carrenza, as the Release app would then stop linking to them.

Therefore, always link to dashboards. I think it's worth having a
dashboard for everything, so this logic shouldn't be used.

1: 5117ce14178d19265fe69a3d967ed7501e2af0d3