Make the Gemfile less specific
Specify less about the versions in the Gemfile. This should reduce the
diff size in Dependabot Pull Requests, and make the Gemfile generally
clearer to read. It's fine specifying a very specific version in the
Gemfile, but only if there's a specific reason which is also
mentioned.

The full version information is available in the Gemfile.lock