Add helper method to create application node breadcrumb
Add an application node breadcrumb helper which when given an
application instance will return a hash with a title and url in the form
that the breadcrumb component expects.

Add a test for the new method and also update the test helper to include
DatabaseCleaner setup and teardown code for ActionView::TestCase.

We will use this to dry up the code a bit in subsequent commits.