allow same repo to be used by multiple apps
This is required to record correctly licensify deployments after
licensify migration to AWS where govuk-app-deployment will be used to
deploy the 3 licensify apps (licensify, licensify-admin and
licensify-feed) individually rather than deploying licensify in a single
deployment using the deprecated alphagov-deployment.

In order to achieve that we:
1. remove the need for repo to be unique in the controller and schema
since the 3 licensify apps share the same repo.

2. add the option in the notification api logic to identify an
app by its name and repo pair by using the application_by_name=true
parameter option when making the request to the release app.  By
default, this parameter option is not set or is set to false, we do the
previous behaviour which is to identify an app by its repo

This change requires the change in
[pr](https://github.com/alphagov/govuk-app-deployment/pull/334)