fix the search for app records by application name.
When search for app records by application name, we should use the
normalized app name because the app record is created with the
normalizes app name.

In addition, we should only use the application
name to search because it is unique and there is no need to use the repo
name.