Create app-specific tabs component
The tabs component in the gem is only designed to be used for switching between content on the same page, rather than different URLs.

This follows the same pattern used in Collections Publisher, where we create a new app-specific tabs component, but pull in the govuk-tabs styles from GOVUK Frontend.