Add instructions to check smokey
This adds an instrution to check Smokey is passing in the 'preceding'
environment before deploy in the 'following' environment e.g. one should
check integration before deploying to staging.

In order to get the correct URL for the environment and provider, this
DRYs-up the code in the ApplicationHelper and exposes a new method to
get the correct 'govuk_domain_suffix'.