Add API controller for Applications
This adds a API endpoint to get information about applications as a JSON
response. This creates a new namespace `/api` to contain all API paths
and a new Api::ApplicationController for API controllers. This removes
controller-wide methods not relevent to the APIs (e.g.
protect_from_forgery) and allows API controllers to inherit from
ActiveController::API, which removes some of the inbuilt Rails overhead
when rendering normal Views.