Add application header partial
This partial DRY's up the application views for show, edit, deployments
and stats as they all have the same layout of header name with tab menu.
It also removes the extra breadcrumb for each tab view, as implemetation
with a partial adds extra complexity and the functionality isn't needed
as the tab view provides the navigation.