Clarify the behaviour of a deploy freeze
While the naming of this flag is still accurate, the "Do not deploy"
badge was imprecise. Since we generally write an explanatory note for
a deploy freeze, this renames the badge to complement the note by
calling out the behavioural change, together with a hint about this
for someone changing the setting.