Fix Smokey URL for Integration environment
Our domain suffix rules aren't as clear cut as perhaps originally
intended when this code was written. Integration doesn't use the
`govuk.digital` domain. Consequently, the 'Integration smokey'
link has not worked for a long time, and developers have to then
navigate the Jenkins UI to find it.

This commit fixes that, and restructures the `govuk_domain_suffix`
to be easier to read.