Add a default branch column to Application
This is intended to support the transition from repositories using a git
branch of "master" branch to a "main" branch. To achieve this I have
added a text field that stores the name of the default branch, with a
default value of "master" to apply this to all existing Application
records.