Add default branch to the application form
This is set as a select field as we have only two valid values. I
initially wanted to do this with a radio button however this was rather
jarring visually as you can only seem to use the radio button component
with a heading for the legend.

Unfortunately the select input doesn't support a hint field as per the
input component so this couldn't be consistent with the other fields.