Show errors when there is a problem for site settings
Prior to this change when you submitted a site settings form with an
error you'd not have the error shown. This is functionality that used to
be in this project but I expect it was accidentally removed in https://github.com/alphagov/release/pull/501

I've also updated the error responses to return a 422 status code.