Show errors for unsuccessful deployment form submissions
Prior to this change when you had errors in a deployment form you'd be
just given one vague error of "Failed to create deployment" and lose
your input. This is functionality that used to
be in this project but I expect it was accidentally removed in
https://github.com/alphagov/release/pull/501. This also re-instates
default values for fields that weren't working correctly.

I've also updated the error response to return a 422 status code.