Show errors for unsuccessful application form submissions
Prior to this change when you had errors in an application form you'd be
given all the errors in a flash message and lose
your input. This is functionality that used to
be in this project but I expect it was accidentally removed in
https://github.com/alphagov/release/pull/501. I've also updated the error
response to return a 422 status code.

This also fixes the behaviour for a successful form submission which
would render the show view without any attributes assigned, so it'd be
somewhat bare. This has now been changed to a redirect as per Rails
convention.