Simplify flash handling
We now only have notice flash alerts and these no longer are set as
hashes.

The previous implementation of this had problems with flashes that
weren't `flash.now` as it expected a flashes' keys to be symbols,
however they become strings when accessed from a session (since they are
encoded as JSON).