Improve deployment monitoring on Integration and Production
The previous Deployment screen was missing a link to the Integration
dashboard, and to the Production dashboard. This discouraged devs
from visiting these dashboards.
In the case of the former, this means developers may not see issues
with their deploy until they release to Staging and check Grafana.
In the case of the latter, this means developers may not spot any
issues that only affect Production.

Before:

![before](https://user-images.githubusercontent.com/5111927/109946419-1f978e80-7cd0-11eb-8c65-f7604eabfe6e.png)

After:

![after](https://user-images.githubusercontent.com/5111927/109946442-26260600-7cd0-11eb-885e-eb3df9fdd937.png)