Clarify Smokey needs to be run manually now
In response to: [^1][^2].

This is slightly misleading as Puppet deploys do still run Smokey
automatically, but the deployer should still make the effort to
check this so the language is still appropriate.

The Production checks are arguably unnecessary for apps given the
checks done in previous environments. However, we may as well keep
recommending them for manually deployed apps, which aren't tested
as thoroughly (yet) - and definitely for Puppet.

[^1]: https://github.com/alphagov/govuk-puppet/pull/11611#pullrequestreview-937766486
[^2]: https://github.com/alphagov/govuk-puppet/pull/11611#pullrequestreview-931760518