Switch side of bagder image to improve accessibility
Swapping the side of the badger image on release pages to improve
readability for screen magnifiers. User accessibility testing for other
GOV.UK interfaces has shown that screen magnifier user find it difficult
when they have to scroll left to right to follow content. Having the
badge image on the left, means users may miss the deployment buttons to
right of it.