Rename release page title to reference release
This updates the html title, h1 heading and breadcrumbs to reference the
release. This is more consistent as the page represents the release,
changes associated with that release and what actions can be taken for a
release. This make more logic sense as you don't just deploy an
application, specifically you deploy a release of an application. This
should also help screen reader users better understand the navigation of
application as you can better tell the difference between an application
page and a release page.