Simplify logic to showing deployed_to column
This simplifies the logic so that we check if any of the latest
deployments match a particular tag in that row of the table. Rather than
needing pre-compute the all deployments for a particular tag, then check
if the tag is in the row.

This will also make it easier to extend the logic to match deployments
on other attributes other than tags. This is set towards enabling
deployments from EKS to be shown in Release, as they don't have
information about associated tags, only the assoicated commit.