Remove redundant test
This is the last lingering reference to Carrenza. As far as I can
tell, it tests that the latest deploy for a given environment is
returned, irrespective of the hosting environment (i.e. if we have
`release_123` on AWS and `release_124` on Carrenza, the latter
would be the one returned).

There appears to be no value in this test now (it essentially
duplicates the test [at the top of the file][1]).
Let's just remove it.

[1]: https://github.com/alphagov/release/blob/19fcc72b8697d8aaaacf9354419d5c4d5f83ffa0/test/unit/report/deploy_lag_report_test.rb#L11-L19