Add Repo model
The class uses HTTParty gem to get a list of repositories from
https://docs.publishing.service.gov.uk/repos.json and information about them.
We will need information about what team owns the app and their Slack channel.

The repos.json is currently a single source of truth of app ownership information.

At this stage we are not storing this information in the application's DB as it will
be used by a worker that runs once a day.

HTTParty gem was chosen as it seems the most comonlly used for this purpose
in GOV.UK apps.