Fix application status
“Production and staging not in sync” status was incorrectly reported as
“Undeployed changes in integration”.

This was because `envs[0, 1]` returns only the first element of the array that
is ["production"] or ["production EKS"]. The #in_sync? methods is checking if
the number of unique versions (length of the array) is less or equal to 1 for
given environments. Therefore it could on only ever return true for an array
with one element.

The full array was then compared and since there were version differences, the
status method was returning :undeployed_changes_in_integration

This updates the status method to use #take method which
returns first n elements from the array.  It also adds the tests.