Add a generic Sidekiq worker to post messages to Slack
Adds a Sidekiq worker that uses Slack Incoming Webhook to post messages to
Slack.

The Webhook URL is stored in AWS Secrets Manager and accessible by the Release
app through govuk_helm_charts configuration.

The options validation checks both:
 - that they are strings as simple JSON datatypes have to be used as per
   https://github.com/sidekiq/sidekiq/wiki/Best-Practices#1-make-your-job-parameters-small-and-simple
 - the exact naming

Slack message error will be reported to Sentry so we will be notified, if the
worker failed.