Include non default branch deploys in commit log
The commit log on an application summary page only lists commits
for the default branch (main) and does not include manual deploys
from other branches.

Whilst this is included in the "What's where" section at the bottom
of the page, it's helpful to also include them in the commit log
section where it is easier to see if an environment is currently
out of sync with `main`.

If any environments have not yet been rendered to the table with
its latest deployed commit, we get the last deployed commit for those
environments and add them as final entries to the commit log
table.