Filter Recent Deployments by Environments
This could be useful for things like finding out which deploys have happened on
production recently, to aid in debugging a live incident.

It would also allow going further back in history, as currently the search is
restricted to the last 25 deployments across all envs, so being able to show 25
deployments in a single env will show more results that would otherwise be
lost.