Remove form to record missing deployment
This is rarely used, if ever, as we record deployments automatically and
never need to manaually do this. Removing this feature to save on
maintenance and to simplify the use case for POST "/deploymenents"
endpoint. We use that endpoint also as an API to record deployments from
Argo. Since fixing our CSRF protection, we've broken that endpoint as we
don't include csrf protection token with API requests. If we make it
solely used as an API endpoint, we can remove the need to authenticate
via a session, which make CSRF no longer a attack vector and remove the
need to verify requests.