Get app repository urls from the Dev docs
We now use the Dev Docs as the single source of truth for
information about repositories.

This commit updates various methods so that they are getting the
information from the Dev Docs and avoid referencing information
stored in the database (which will be eventually removed).

If the url is unavailable from the Dev Docs (e.g. if it's a temporary
app or hasn't been added to the Dev Docs apps.json file), the url
is generated by parameterizing the app name.

We are also caching the json data as it does not change too frequently
and avoids unnecessary multiple calls to the Dev docs.