Remove use of GitHub repo path field
The `repo` field stores `alphagov/<repo_name>` and is used to get
information from Github such as commits and deployments. We can
still get this value by using the application name instead, and
is now stored in a new variable `repo_path` which distinguishes
it better from the repo name itself.

As we can get this information from the Developer Docs it is no
longer necessary to include it in the HTML form or be required in
the creation of an Application and has now been removed.