Update deployments to get application by name
The POST create "/deployments" endpoint is used as an API to record
deployments from Argo. It passes the `repo` parameter in the form
of `alphagov/<repo_name>` - see https://github.com/alphagov/govuk-helm-charts/blob/main/charts/argo-services/templates/workflows/notify-release/workflow.yaml#L27

This commit simplifies the `repo_path` method to extract the repo
name which can be used to check if it is an existing app in the
database. Some methods have also been removed that are not being used.