Get `shortname` for applications from the Dev Docs
We can't remove the `shortname` column from the database as it is
used to generate friendly slugs for the application URLs. We can
however change how the fallback shortname is generated (which is
used when one hasn't been specified).

Instead of generating it using the `repo_path` we can now get this
directly from the Developer Docs. Failing this, it will be created
by parameterizing the app nane.