Move "not on default branch" label to deployment table
This simplies the ability for users to see whether a deployment is of a
commit not on the main branch. Previously this was displayed as commits
amended to the commit log, which can be confusing as they aren't
chronologically or related to commit history of the default branch.

This work reduces the number of API calls needed to be made, as we don't
need to display extra information about commit.