Add functionality to delete retired apps
Retired apps are being flagged in the Out of sync release
alert and there is no way to remove them in the Release app.

This commit allows an app to have the option of being deleted on
the Edit page. There is a popup confirmation dialog when the delete
button is pressed to ensure it is not removed by mistake.