Allow alerts to be posted to a different Slack channel
At the moment, out-of-sync alerts are sent to the Slack channel of the
team who own the application.

Some teams may wish to have automated alerting sent to a different
channel. Therefore updating to use the `alerts_team` value instead of
the `team` value.

By default, `govuk-developer-docs` falls back to the `team` value where
no other value is specified.