Fix Sass deprecation warnings
Sass's behaviour for declarations that appear after nested rules is changing,
triggering deprecation warnings. To maintain compatibility with future Sass
versions, this commit refactors the affected selectors in the stylesheets.

The linter threw an "Unexpected duplicate selector '&'"
error when using `& {}` to wrap the declarations. To resolve this, the order of
declarations was adjusted to place them above the mixin (`@include govuk-font(...)`),
which avoids both the deprecation warning and the linter error.