Add kubeclient to serve as the k8s client
Decided to use `kubeclient`. It's well supported, the most popular (44,286k
downloads, compared to 749k download of k8s-ruby), good docs including
specifically for AWS EKS as well as helper method to generate the
authentication token.

On AWS EKS by default the authentication method is IAM. Hence the flow is as
described below:

1. Release makes request to assume a `release-assumed` AWS IAM role in an AWS
account to Security Token Service (STS)
The `kubeclient` gem has the `Kubeclient::AmazonEksCredentials` class which can
accept a set of IAM credentials and contains a helper method to generate the
authentication token.  It requires a set of gems which are not included
in kubeclient dependencies (aws-sigv4). We also require either the aws-sdk v2
or aws-sdk-core v3 gems to generate the required Aws:Credentials object to pass
to this method.

2. STS responds with temporary credentials
AWS STS generates temporary security credentials for the requested role, which
includes an Access Key ID, Secret Access Key, and a Session Token. These
credentials are valid for one minute.

3. Release uses those credentials to generate a K8s access token.
Once "Release has received the temporary AWS credentials from STS, it uses
these credentials to authenticate to a Kubernetes (K8s) cluster.