Update Smart Answers shortname
The shortname was changed in the DB via Rails console form "smartanswers" to
"smart-answers" so we can fetch data from k8s API. This updated other
references in the app. This matches other references to this app/repo

The shortnames are generated using the FriendlyId gem (name.parameterize) so
the new string matches this pattern.