Add change failure notes to deployments
We want to be able to track the number of deployments which contained a change failure so that we can report on DORA metrics.

To enable this, we have created a new singular resource for change failures and added a toggle to applications that will be used to switch the change failure feature on or off per app.